refactor: TokenController 제거
