--- NangPaGo-be/src/main/java/com/mars/NangPaGo/domain/auth/controller/TokenController.java
@@ -1,26 +1,0 @@
-package com.mars.NangPaGo.domain.auth.controller;
-
-import com.mars.NangPaGo.domain.auth.service.TokenService;
-import io.swagger.v3.oas.annotations.tags.Tag;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RequiredArgsConstructor
-@Tag(name = "토큰 API", description = "토큰 관련 API")
-@RequestMapping("/api/token")
-@RestController
-public class TokenController {
-
-    private final TokenService tokenService;
-
-    @PostMapping("/reissue")
-    public ResponseEntity<Void> reissue(HttpServletRequest request, HttpServletResponse response) {
-        tokenService.reissueTokens(request, response);
-        return ResponseEntity.ok().build();
-    }
-}
