+++ NangPaGo-be/src/main/java/com/mars/NangPaGo/domain/user/entity/User.java
@@ -6,14 +6,8 @@ import com.mars.NangPaGo.domain.favorite.recipe.entity.RecipeFavorite;
 import com.mars.NangPaGo.domain.recipe.entity.RecipeLike;
 import com.mars.NangPaGo.domain.user.enums.Gender;
 import com.mars.NangPaGo.auth.enums.OAuth2Provider;
-import jakarta.persistence.CascadeType;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.OneToMany;
+import jakarta.persistence.*;
+
 import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -42,6 +36,7 @@ public class User extends BaseEntity {
     private String role;
 
     @Enumerated(EnumType.STRING)
+    @Column(name = "provider")
     private OAuth2Provider oauth2Provider;
     private String providerId;
